requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.19.3
transformers>=4.30.0
gradio==5.18.0
fastapi>=0.100.0
uvicorn>=0.22.0
Pillow>=9.5.0
opencv-python-headless>=4.7.0
numpy>=1.24.3 